agent/requirements.txt
langchain==0.3.26 
langchain-openai==0.3.25 
langchain-community==0.3.26
langchain-core==0.3.66 
fastmcp==2.8.1
python-dotenv
requests
asyncio
openai-agents